Huge rewrite of GUI code.
ControlTab has been: (1) reorganized, and (2) largely rewritten to use the
GridBagLayout. This has the advantage of being much more flexible than the other
layouts. It might look a bit daunting to change stuff now, but is best in the
long run. Also, the buttons and stuff don't get all whacked when you resize the
window.